Don't silently fail if Fuse can't be started.
subprocess.call() returns an error code that we didn't check. Instead
use subprocess.check_output() that throws an exception if the subprocess
returns non-zero. In that case, log the output from Fuse and alert the
user that there is a problem.

An argument was made that the user doesn't need to be alerted about this
problem, since staring Fuse is not a result of a deliberate user action.
In my opinion it is important for the user to be aware that there is an
error, so he or she can try to fix it. The alternative is that things
like code completion just don't work.