Fix bug in `getExtension()`
This method never worked, as it would return the file name instead of
the extension. However, it was never used for getting file extensions,
but for getting syntax names. So I introduced a new function
`getSyntax()` for that instead.

We can now use `getExtension()` in `FuseFocusDesigner.is_enabled()`, as
suggested in the code review of this branch.