Make FocusEditor work when multiple Sublime windows are open
Previously, a `FocusEditor` request would fail if a window for that
project was open, but was not the active window.

On Windows, the correct window will also be raised to the foreground. On
OS X however, the Sublime app itself (with all its windows) will be
raised, leaving the previously topmost window at the top. There is no way to
raise a specific Sublime window to the front on OS X, see for instance
the discussion on
https://forum.sublimetext.com/t/how-to-focus-window/5219.